refactor: move commands and filter query hover into a seperate folders
